requirements.txt
Django==1.8
psycopg2==2.6.2
pillow==3.0.0
djangorestframework==3.5.4